Added Reporting and Notifications Features, Invitation Features, and more, Testing Pending.

+++ backend/threaddit/comments/models.py
@@ -1,17 +1,22 @@
 from datetime import UTC, datetime
 from typing import TYPE_CHECKING
 
+import sqlalchemy as sa
+from flask import abort
+from flask_jwt_extended import current_user
 from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.schema import ForeignKey
 
 from threaddit import db
 from threaddit.media.models import Media, OpType
+from threaddit.notifications.models import Notifications, NotifType
+from threaddit.reactions.models import Reactions
+from threaddit.saves.models import Saves
 
 if TYPE_CHECKING:
     from threaddit.comments.schemas import CommentBodyType
     from threaddit.posts.models import Posts
-    from threaddit.reactions.models import Reactions
     from threaddit.threads.models import Thread
     from threaddit.users.models import User
 
@@ -28,15 +33,35 @@ class Comments(db.Model):
     created_at: Mapped[datetime] = mapped_column(default=db.func.now(), nullable=False)
     upvotes: Mapped[int] = mapped_column(default=0, nullable=False)
     downvotes: Mapped[int] = mapped_column(default=0, nullable=False)
+    depth: Mapped[int] = mapped_column(nullable=False, default=0)
     saved_count: Mapped[int] = mapped_column(default=0)
+    report_count: Mapped[int] = mapped_column(default=0)
     is_deleted: Mapped[bool] = mapped_column(default=False, nullable=False)
+    is_removed: Mapped[bool] = mapped_column(default=False, nullable=False)
+    is_locked: Mapped[bool] = mapped_column(default=False, nullable=False)
+    is_sticky: Mapped[bool] = mapped_column(default=False, nullable=False)
     replies_count: Mapped[int] = mapped_column(default=0, nullable=False)
+    disable_reports: Mapped[bool] = mapped_column(default=False, nullable=False)
     reaction: Mapped[list["Reactions"]] = relationship(back_populates="comment")
     media: Mapped["Media"] = relationship()
     user: Mapped["User"] = relationship(back_populates="comments")
     post: Mapped["Posts"] = relationship(back_populates="comment")
     parent: Mapped["Comments"] = relationship(back_populates="children", remote_side="Comments.id")
     children: Mapped[list["Comments"]] = relationship(back_populates="parent", foreign_keys="Comments.parent_id")
+
+    @property
+    def has_upvoted(self):
+        reaction = db.session.scalar(
+            sa.select(Reactions).where(Reactions.user_id == current_user.id, Reactions.comment_id == self.id)
+        )
+        return reaction.is_upvote if reaction else None
+
+    @property
+    def has_saved(self):
+        saved_comment = db.session.scalar(
+            sa.select(Saves).where(Saves.user_id == current_user.id, Saves.comment_id == self.id)
+        )
+        return saved_comment is not None
 
     @hybrid_property
     def karma(self) -> int:
@@ -56,14 +81,13 @@ class Comments(db.Model):
     def vote_ratio(cls):
         return cls.upvotes / (cls.upvotes + cls.downvotes)
 
-    @classmethod
+    @staticmethod
     def add(
-        cls,
         form: "CommentBodyType",
         thread: "Thread",
         user: "User",
         post: "Posts",
-        comment: "Comments | None" = None,
+        comment: "Comments|None" = None,
     ):
         new_comment = Comments(
             user_id=user.id,
@@ -73,12 +97,14 @@ class Comments(db.Model):
         if comment:
             comment.replies_count += 1
             new_comment.parent_id = comment.id
+            new_comment.depth = comment.depth + 1
         if form["media"]:
             new_comment.media = Media.add(f"comments/{new_comment.id}", form=form)
         post.comment_count += 1
         user.comment_count += 1
         thread.comment_count += 1
         db.session.add(new_comment)
+        new_comment.notify_user(user, thread, post, comment)
         return new_comment
 
     # noinspection DuplicatedCode
@@ -102,13 +128,50 @@ class Comments(db.Model):
             user.comment_count -= 1
             if self.parent:
                 self.parent.replies_count -= 1
-            self.media.delete()
+            if self.media:
+                self.media.delete()
             # noinspection PyTypeChecker
             self.is_deleted = True
         if post_delete:
             for reaction in self.reaction:
                 reaction.remove()
             db.session.delete(self)
+
+    def notify_user(
+        self,
+        user: "User",
+        thread: "Thread",
+        post: "Posts",
+        comment: "Comments|None" = None,
+    ):
+        if comment and comment.user_id == user.id or post.user_id == user.id:
+            return
+        if comment:
+            Notifications.notify(
+                NotifType.COMMENT_REPLY,
+                user=comment.user,
+                title=f"{user.username} replied to your comment on {thread.name}",
+                sub_title=comment.content,
+                content=self.content,
+                res_id=post.id,
+                res_media_id=user.avatar_id,
+            )
+        else:
+            Notifications.notify(
+                NotifType.POST_COMMENT,
+                user=post.user,
+                title=f"{user.username} commented on your post on {thread.name}",
+                sub_title=post.title,
+                content=self.content,
+                res_id=post.id,
+                res_media_id=user.avatar_id,
+            )
+
+    def validate_comment(self, thread: "Thread", post: "Posts"):
+        post.validate_post(thread)
+        if self.post_id != post.id:
+            return abort(400, {"message": f"Comment does not belong to post {post.id}"})
+        return None
 
     # noinspection PyTypeChecker
     def __init__(self, user_id: int, content: str, post_id: int):
